+++ codex/signals/os_signals.py
@@ -3,6 +3,7 @@
 import asyncio
 import signal
 from asyncio import Event
+from sys import platform
 
 from loguru import logger
 
@@ -42,10 +43,14 @@ def _restart_signal_handler(*_args):
 def bind_signals_to_loop():
     """Binds signals to the handlers."""
     try:
-        loop = asyncio.get_running_loop()
+        if platform == "win32":
+            sig_add = signal.signal
+        else:
+            loop = asyncio.get_running_loop()
+            sig_add = loop.add_signal_handler
         for name in STOP_SIGNAL_NAMES:
             if sig := getattr(signal, name, None):
-                loop.add_signal_handler(sig, _shutdown_signal_handler)
-        loop.add_signal_handler(signal.SIGUSR1, _restart_signal_handler)
+                sig_add(sig, _shutdown_signal_handler)
+        sig_add(signal.SIGUSR1, _restart_signal_handler)
     except NotImplementedError:
         logger.info("Shutdown and restart signal handling not implemented on windows.")