+++ eslint.config.js
@@ -28,17 +28,17 @@ export const FLAT_RECOMMENDED = "flat/recommended";
 
 export const CONFIGS = {
   js: {
-    ...eslintJs.configs.recommended,
+    ...eslintJs.configs.all,
     ...eslintPluginArrayFunc.configs.all,
     ...eslintPluginComments.recommended,
     ...eslintPluginCompat.configs[FLAT_RECOMMENDED],
     ...eslintPluginDepend.configs[FLAT_RECOMMENDED],
-    ...eslintPluginImport.flatConfigs.recommended,
+    ...eslintPluginImport.flatConfigs.all,
     ...eslintPluginNoUnsanitized.configs.recommended,
-    ...eslintPluginPromise.configs[FLAT_RECOMMENDED],
-    ...eslintPluginRegexp.configs[FLAT_RECOMMENDED],
-    ...eslintPluginSonarjs.configs.recommended,
-    ...eslintPluginUnicorn.configs.recommended,
+    ...eslintPluginPromise.configs[FLAT_ALL],
+    ...eslintPluginRegexp.configs[FLAT_ALL],
+    ...eslintPluginSonarjs.configs.all,
+    ...eslintPluginUnicorn.configs.all,
     plugins: {
       depend: eslintPluginDepend,
       "no-secrets": eslintPluginNoSecrets,
@@ -47,12 +47,11 @@ export const CONFIGS = {
       unicorn: eslintPluginUnicorn,
     },
     languageOptions: {
-      // eslint-plugin-import sets this to 2018.
+      // eslint-plugin-import recommended sets this to 2018.
       ecmaVersion: "latest",
     },
     rules: {
-      "@stylistic/multiline-comment-style": "off", // https://github.com/eslint-stylistic/eslint-stylistic/issues/968
-      "array-func/prefer-array-from": "off", // for modern browsers the spread operator, as preferred by unicorn, works fine.
+      "@stylistic/multiline-comment-style": "off", // Multiple bugs with this rule
       "max-params": ["warn", 4],
       "no-console": "warn",
       "no-debugger": "warn",
@@ -60,14 +59,6 @@ export const CONFIGS = {
       "security/detect-object-injection": "off",
       "simple-import-sort/exports": "warn",
       "simple-import-sort/imports": "warn",
-      "space-before-function-paren": "off",
-      "unicorn/filename-case": [
-        "error",
-        { case: "kebabCase", ignore: [".*.md"] },
-      ],
-      "unicorn/prefer-node-protocol": "off",
-      "unicorn/prevent-abbreviations": "off",
-      "unicorn/switch-case-braces": ["warn", "avoid"],
     },
   },
 };
@@ -100,9 +91,9 @@ export default defineConfig([
       "uv.lock",
     ],
   },
-  eslintPluginPrettierRecommended,
   eslintPluginSecurity.configs.recommended,
   eslintPluginStylistic.configs.all,
+  eslintPluginPrettierRecommended,
   {
     languageOptions: {
       globals: {