+++ NEWS.md
@@ -6,6 +6,12 @@ width: 128px;
 border-radius: 128px;
 " />
 
+## v1.8.14
+
+- Fixes
+    - Fix settings not saving in reader and sometimes in browser.
+    - Fix shutdown and restart signals on Windows.
+
 ## v1.8.13
 
 - Fixes